+++ examples/dependencies/child/flake.nix
@@ -11,7 +11,7 @@
       system = "x86_64-linux";
 
       pkgs' = pkgs.legacyPackages."${system}";
-      nedryland = import ../../default.nix;
+      nedryland = import ../../../default.nix;
       project = import ./project.nix {
         inherit nedryland;
         pkgs = pkgs';
@@ -19,7 +19,7 @@
     in
     {
       packages."${system}" = project.matrix // {
-        default = pkgs'.linkFarmFromDrvs "all" (pkgs'.lib.flatten project.matrix.all);
+        default = project.all;
       };
       devShells."${system}" = project.shells;
     };