+++ examples/hello/libraries/greeting/greeting.nix
@@ -1,13 +1,15 @@
 { base, writeScriptBin }:
 base.mkLibrary rec{
-  name = "greetinglib";
+  name = "greetingLib";
   version = "1.0.0";
+
   lib = writeScriptBin name ''
     case $1 in
       swedish) echo "hej";;
       english) echo "hello";;
       french) echo "bonjour";;
-      finish) echo "hyvää päivää";;
+      finnish) echo "hyvää päivää";;
+      finish) echo "🏁";;
       *) >&2 echo "Unknown language $1"; exit 1;;
     esac
   '';