+++ test.nix
@@ -2,7 +2,7 @@
 let
   nedryland = import ./default.nix;
   # add any examples you want tested here
-  tests = {
+  examples = {
     hello = import ./examples/hello/project.nix {
       inherit pkgs nedryland;
     };
@@ -20,10 +20,11 @@ let
         project.override { enableChecks = true; }
       ).all
     )
-    tests) // {
-    combinedDeployment = builtins.trace "🎪 Running combined deployment test." import ./test/deployment.nix tests.hello.matrix.combinedDeployment.deploy pkgs.lib.assertMsg;
-    sameWhenFiltered = builtins.trace "🎡 Running filter source tests." import ./test/filter-source.nix pkgs.lib.assertMsg tests.hello.matrix.sameThing1 tests.hello.matrix.sameThing2;
-    docsTest = builtins.trace "🦖 Running docs tests." import ./test/docs.nix pkgs tests.documentation.matrix;
+    examples) // {
+    combinedDeployment = builtins.trace "🎪 Running combined deployment test." import ./test/deployment.nix examples.hello.matrix.combinedDeployment.deploy pkgs.lib.assertMsg;
+    sameWhenFiltered = builtins.trace "🎡 Running filter source tests." import ./test/filter-source.nix pkgs.lib.assertMsg examples.hello.matrix.sameThing1 examples.hello.matrix.sameThing2;
+    docsTest = builtins.trace "🦖 Running docs tests." import ./test/docs.nix pkgs examples.documentation.matrix;
+    componentFnsTest = builtins.trace "📦 Running componentFns tests." import ./test/components.nix pkgs.lib.assertMsg;
   };
 in
 (mappedTests // {